+++ modules/http_requests/__init__.py
@@ -1,0 +1,6 @@
+from .http_request import http_request
+from .http_user_agent_get import http_user_agent_get
+from .http_accept_language_get import http_accept_language_get
+from .http_user_agent_post import http_user_agent_post
+from .http_accept_language_post import http_accept_language_post
+from .http_aizawa_ninja import http_aizawa_ninja
