+++ modules/utils/colors.py
@@ -1,0 +1,9 @@
+# Colors
+RED = "\033[31m"
+BOLD = "\033[1m"
+CYAN = "\033[36m"
+BLUE = "\033[34m"
+CLEAR = "\033[0m"
+GREEN = "\033[32m"
+PURPLE = "\033[35m"
+YELLOW = "\033[33m"
